+++ aiogram_dialog/api/protocols/__init__.py
@@ -1,11 +1,13 @@
 __all__ = [
+    "DialogProtocol",
     "ManagedDialogProtocol",
     "ActiveDialogManager", "BaseDialogManager",
     "MediaIdStorageProtocol",
     "DialogProtocol", "DialogRegistryProtocol",
 ]
 
+from .dialog import DialogProtocol
 from .managed import ManagedDialogProtocol
 from .manager import ActiveDialogManager, BaseDialogManager
 from .media import MediaIdStorageProtocol
-from .registry import DialogProtocol, DialogRegistryProtocol
+from .registry import DialogRegistryProtocol
